Started on GPU and Parsl support for Freesurfer step

+++ s2a_bedpostx.py
@@ -29,7 +29,7 @@ if not exists(FSLDIR):
 bedpostx = join(odir,"bedpostx_b1000")
 bedpostxResults = join(odir,"bedpostx_b1000.bedpostX")
 EDI = join(odir, "EDI")
-"""
+
 smart_copy(join(odir,"data_eddy.nii.gz"),join(bedpostx,"data.nii.gz"),args.force)
 smart_copy(join(odir,"data_bet_mask.nii.gz"),join(bedpostx,"nodif_brain_mask.nii.gz"),args.force)
 smart_copy(join(odir,"bvals"),join(bedpostx,"bvals"),args.force)
@@ -57,11 +57,14 @@ if args.force or not exists(join(bedpostxResults, "dyads2_dispersion.nii.gz")):
                                                  join(bedpostxResults, "merged_ph2samples"),
                                                  join(bedpostxResults, "nodif_brain_mask"),
                                                  join(bedpostxResults, "dyads2")))
-"""
+
 if args.force or not exists(join(EDI, "bedpostx_b1000.bedpostX")):
     if exists(join(EDI, "bedpostx_b1000.bedpostX")):
         rmtree(join(EDI, "bedpostx_b1000.bedpostX"))
-    copytree(bedpostxResults, join(EDI, "bedpostx_b1000.bedpostX"))
+    if exists(bedpostxResults):
+        copytree(bedpostxResults, join(EDI, "bedpostx_b1000.bedpostX"))
+    else:
+        print("Failed to generate bedpostX results")
   
 if args.output_time:
     printFinish(start_time)
